requirements.txt
asgiref==3.8.1
Django==5.1.6
django-filter==25.1
django-quill-editor==0.1.42
djangorestframework==3.15.2
pillow==11.1.0
sqlparse==0.5.3
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
